# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elevator Inspection Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in organization (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/invite:
    post:
      operationId: inviteUser
      tags: [users]
      summary: Invite a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserBody"
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
        isActive:
          type: boolean
        organizationId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - isActive
        - organizationId
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [ADMIN, USER]
        isActive:
          type: boolean
    InviteUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, USER]
        password:
          type: string
      required:
        - email
        - role
        - password
...